+++ packages/ethereum-contracts/test/foundry/superfluid/Superfluid.t.sol
@@ -6,8 +6,11 @@ import { UUPSProxiable } from "../../../contracts/upgradability/UUPSProxiable.so
 import { SuperToken } from "../../../contracts/superfluid/SuperToken.sol";
 import { SuperTokenV1Library } from "../../../contracts/apps/SuperTokenV1Library.sol";
 import { ISuperAgreement } from "../../../contracts/interfaces/superfluid/ISuperAgreement.sol";
-import { ISuperfluid } from "../../../contracts/interfaces/superfluid/ISuperfluid.sol";
+import { ISuperfluid, SuperAppDefinitions } from "../../../contracts/interfaces/superfluid/ISuperfluid.sol";
+import { ISuperApp } from "../../../contracts/interfaces/superfluid/ISuperApp.sol";
 import { AgreementMock } from "../../../contracts/mocks/AgreementMock.t.sol";
+import { SuperAppMockNotSelfRegistering } from "../../../contracts/mocks/SuperAppMocks.t.sol";
+import { SimpleACL } from "../../../contracts/utils/SimpleACL.sol";
 
 contract SuperfluidIntegrationTest is FoundrySuperfluidTester {
     using SuperTokenV1Library for SuperToken;
@@ -31,7 +34,9 @@ contract SuperfluidIntegrationTest is FoundrySuperfluidTester {
             vm.startPrank(sf.governance.owner());
             sf.governance.registerAgreementClass(sf.host, address(agreementMock));
             vm.stopPrank();
-            agreementMock = sf.host.NON_UPGRADABLE_DEPLOYMENT() ? agreementMock : AgreementMock(address(sf.host.getAgreementClass(id)));
+            agreementMock = sf.host.NON_UPGRADABLE_DEPLOYMENT() 
+                ? agreementMock 
+                : AgreementMock(address(sf.host.getAgreementClass(id)));
             mocks[i + _NUM_AGREEMENTS] = ISuperAgreement(address(agreementMock));
         }
 
@@ -81,6 +86,68 @@ contract SuperfluidIntegrationTest is FoundrySuperfluidTester {
         vm.startPrank(newAdmin);
         vm.expectRevert(ISuperfluid.HOST_ONLY_GOVERNANCE.selector);
         sf.host.changeSuperTokenAdmin(superToken, newAdmin);
+        vm.stopPrank();
+    }
+
+    function testSuperAppRegistrationViaSimpleACL() public {
+        SimpleACL simpleAcl = new SimpleACL();
+        Superfluid hostWithSimpleACL = new Superfluid(
+            true, true, 3_000_000, address(0), address(0), address(simpleAcl)
+        );
+        ISuperApp mockSuperApp1 = ISuperApp(address(new SuperAppMockNotSelfRegistering()));
+        ISuperApp mockSuperApp2 = ISuperApp(address(new SuperAppMockNotSelfRegistering()));
+
+        hostWithSimpleACL.initialize(sf.governance);
+
+        bytes32 aclSuperAppRegRole = hostWithSimpleACL.ACL_SUPERAPP_REGISTRATION_ROLE();
+
+        // first, give permission to alice
+        simpleAcl.grantRole(aclSuperAppRegRole, alice);
+
+        // as bob, try to register a superapp - should revert
+        vm.startPrank(bob);
+        vm.expectRevert();
+        hostWithSimpleACL.registerApp(mockSuperApp1, SuperAppDefinitions.APP_LEVEL_FINAL);
+
+        // as alice, try to register a superapp - should succeed
+        vm.startPrank(alice);
+        hostWithSimpleACL.registerApp(mockSuperApp1, SuperAppDefinitions.APP_LEVEL_FINAL);
+        vm.stopPrank();
+        vm.assertTrue(hostWithSimpleACL.isApp(mockSuperApp1));
+
+        // revoke permission from alice
+        simpleAcl.revokeRole(aclSuperAppRegRole, alice);
+
+        // as alice, try to register a superapp - should revert
+        vm.startPrank(alice);
+        vm.expectRevert();
+        hostWithSimpleACL.registerApp(mockSuperApp2, SuperAppDefinitions.APP_LEVEL_FINAL);
+        vm.stopPrank();
+
+        // nobody else can grant permission to register superapps
+        vm.startPrank(eve);
+        vm.expectRevert();
+        simpleAcl.grantRole(aclSuperAppRegRole, bob);
+        vm.stopPrank();
+
+        // nobody can define admin roles ...
+        bytes32 dedicatedAdminRole = keccak256("SUPER_APP_REGISTRATION_ADMIN");
+        vm.startPrank(eve);
+        vm.expectRevert();
+        simpleAcl.setRoleAdmin(aclSuperAppRegRole, dedicatedAdminRole);
+        vm.stopPrank();
+
+        // ... except the default admin
+        // (This is to be done in a possible future where we start using this ACL for other purposes too
+        // and want to have a more sophisticated permissioning scheme.)
+        simpleAcl.setRoleAdmin(aclSuperAppRegRole, dedicatedAdminRole);
+
+        // grant heidi the new admin role
+        simpleAcl.grantRole(dedicatedAdminRole, heidi);
+
+        // now heidi can manage permissions for superapp deployers
+        vm.startPrank(heidi);
+        simpleAcl.grantRole(aclSuperAppRegRole, bob);
         vm.stopPrank();
     }
 }